requirements.txt
asgiref==3.7.2
Django==4.2.4
django-environ==0.10.0
django-jazzmin==2.6.0
Pillow==10.0.0
psycopg2-binary==2.9.7
python-dateutil==2.8.2
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
